a snarled hairball of bad data
Not sure what is up, but some docs fly by, other docs get stuck on the
way out.  CouchDB will hang up, fill RAM, and crash trying to spit out
some docs.  While this is dangerous, the only way to move on is to try
to get these docs, crash, then move on to the next doc.